修复了不能加载多个纹理等bug

+++ OpenGL/OpenGL/OpenGL/src/tests/TestClearColor.cpp
@@ -15,7 +15,7 @@ namespace test {
 
 	}
 
-	void TestClearColor::OnUpdate(float deltaTime)
+	void TestClearColor::OnUpdate(double deltaTime)
 	{
 
 	}
